Address #CA1816 warning in implement disposeasync article
