+++ docs/standard/garbage-collection/snippets/dispose-async/AnotherAsyncDisposable.cs
@@ -1,9 +1,9 @@
-using System;
+﻿using System;
 using System.Threading.Tasks;
 
 public class AnotherAsyncDisposable : IAsyncDisposable
 {
     public AnotherAsyncDisposable() => throw new Exception("Oops, sorry...");
 
-    public ValueTask DisposeAsync() => new ValueTask();
+    public ValueTask DisposeAsync() => new();
 }
