+++ docs/standard/garbage-collection/snippets/dispose-async/ExampleConfigureAwaitProgram.cs
