+++ docs/standard/garbage-collection/snippets/dispose-async/ExamplePatterns.cs
