+++ docs/standard/garbage-collection/snippets/dispose-async/ExampleUsingDeclarationProgram.cs
