+++ docs/standard/garbage-collection/snippets/dispose-async/dispose-async.csproj
@@ -1,9 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
-    <StartupObject>ExampleUsingDeclarationProgram</StartupObject>
+    <StartupObject>ExampleUsingStatementProgram</StartupObject>
   </PropertyGroup>
 
 </Project>
